All your UART implementation puted to usart.c file and public functions declarations in usart.h file
