+++ Lecture5_c/UART.uvprojx
@@ -402,6 +402,11 @@
               <FileType>1</FileType>
               <FilePath>.\STM32F30x_StdPeriph_Driver\src\stm32f30x_usart.c</FilePath>
             </File>
+            <File>
+              <FileName>USART.c</FileName>
+              <FileType>1</FileType>
+              <FilePath>.\Applcation\USART.c</FilePath>
+            </File>
           </Files>
         </Group>
         <Group>
